// KentOS.Kalem.Web/frontend/src/data/types.generated.ts
export interface components {
    schemas: {
// ...
        IstatistikDilimDto: {
            etiket?: string | null;
            /** Format: int32 */
            deger?: number;
            /** Format: double */
            yuzde?: number;
            renk?: string | null;
        };
// ...
        IstatistikSeriNoktasiDto: {
            etiket?: string | null;
            tarih?: string | null;
            /** Format: int32 */
            deger?: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}